+++ Telemonitor/main.py
@@ -33,10 +33,24 @@ def run():
                 reply_markup=ikb.get_keyboard()
             )
 
+    if cfg["enable_file_transfer"]:
+        @dp.message_handler(content_types=['document', 'photo'])
+        async def __file_transfer(message: types.Message):
+            if TM_Whitelist.is_whitelisted(message.from_user.id):
+                h.init_shared_dir()
+                if message.content_type == 'document':
+                    await message.document.download(path.join(h.PATH_SHARED_DIR, message.document.file_name))
+                    logger.info(f"Successfully downloaded file {message.document.file_name} to {path.abspath(h.PATH_SHARED_DIR)}")
+                    await message.reply(text=f"Successfully downloaded file {message.document.file_name}", reply=False)
+                elif message.content_type == 'photo':
+                    await message.photo[-1].download(h.PATH_SHARED_DIR)
+                    logger.info(f"Successfully downloaded image(-s) to {path.join(path.abspath(h.PATH_SHARED_DIR), 'photos')}")
+                    await message.reply(text="Successfully downloaded image(-s)", reply=False)
+
     print(f'Bot is starting. Version: {__version__}')
     executor.start_polling(
         dp,
         skip_updates=True,
-        on_startup=lambda _: TM_Whitelist.send_to_all(bot, code("System was booted")),
-        on_shutdown=lambda _: TM_Whitelist.send_to_all(bot, code("System is shutting down"))
+        on_startup=(lambda _: TM_Whitelist.send_to_all(bot, code("System was booted"))) if cfg["state_notifications"] else None,
+        on_shutdown=(lambda _: TM_Whitelist.send_to_all(bot, code("System is shutting down"))) if cfg["state_notifications"] else None
     )