+++ Telemonitor/main.py
@@ -49,11 +49,11 @@ def run():
                 h.init_shared_dir()
                 if message.content_type == 'document':
                     await message.document.download(path.join(h.PATH_SHARED_DIR, message.document.file_name))
-                    logger.info(f"Successfully downloaded file '{message.document.file_name}' to '{path.abspath(h.PATH_SHARED_DIR)}'")
+                    logger.info(f'Successfully downloaded file "{message.document.file_name}" to "{path.abspath(h.PATH_SHARED_DIR)}""')
                     await message.reply(text=f"Successfully downloaded file {code(message.document.file_name)}", parse_mode=PARSE_MODE, reply=False)
                 elif message.content_type == 'photo':
                     await message.photo[-1].download(h.PATH_SHARED_DIR)
-                    logger.info(f"Successfully downloaded image(-s) to {path.join(path.abspath(h.PATH_SHARED_DIR), 'photos')}")
+                    logger.info(f'Successfully downloaded image(-s) to "{path.join(path.abspath(h.PATH_SHARED_DIR), "photos")}"')
                     await message.reply(text="Successfully downloaded image(-s)", parse_mode=PARSE_MODE, reply=False)
 
     print(f'Bot is starting. Version: {__version__}')