+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Telemonitor Changelog
 
+[**Unreleased**](https://github.com/maximilionus/Telemonitor)
+- Enhanced already imported modules usage
+
 [**0.2.5**](https://github.com/maximilionus/Telemonitor/releases/tag/v0.2.5)
 - Added automatic logs cleanup after exceeding the log files limit. Look for const `MAX_LOGS` in [`helpers.py`](./Telemonitor/helpers.py)
 - Added configuration file loading optimization (Read config file values only if this file was modified)