+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added simple unit test for script version match
 - Enhanced already imported modules usage
 - Logging level changed to `INFO`
+- Added bot version to logging
 
 [**0.2.5**](https://github.com/maximilionus/Telemonitor/releases/tag/v0.2.5)
 - Added automatic logs cleanup after exceeding the log files limit. Look for const `MAX_LOGS` in [`helpers.py`](./Telemonitor/helpers.py)